unify bitpack files
